+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-08-23 18:33:24
-// Commit: fix(ui): prevent double tap duplicate action in ui
+// Updated at: 2026-08-24 10:32:13
+// Commit: perf(ui): apply RepaintBoundary on animated ui widgets
 
 class uiComponentState {
-  final String commitRef = '3853';
-  final String lastMessage = 'fix(ui): prevent double tap duplicate action in ui';
-  final DateTime timestamp = DateTime.parse('2026-08-23T18:33:24+05:00');
+  final String commitRef = '3871';
+  final String lastMessage = 'perf(ui): apply RepaintBoundary on animated ui widgets';
+  final DateTime timestamp = DateTime.parse('2026-08-24T10:32:13+05:00');
 }